+++ engine/flink/avro-util/src/test/scala/pl/touk/nussknacker/engine/avro/source/KafkaAvroSourceFactorySpec.scala
@@ -2,21 +2,35 @@ package pl.touk.nussknacker.engine.avro.source
 
 import java.nio.charset.StandardCharsets
 
+import com.typesafe.config.ConfigFactory
 import io.confluent.kafka.schemaregistry.client.{SchemaRegistryClient => CSchemaRegistryClient}
 import org.apache.avro.Schema
 import org.apache.avro.generic.GenericRecord
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.api.scala._
 import org.scalatest.Assertion
+import pl.touk.nussknacker.engine.Interpreter
+import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.{CustomNodeError, NodeId}
+import pl.touk.nussknacker.engine.api.context.ValidationContext
 import pl.touk.nussknacker.engine.api.context.transformation.TypedNodeDependencyValue
 import pl.touk.nussknacker.engine.api.process.{Source, TestDataGenerator, TestDataParserProvider}
 import pl.touk.nussknacker.engine.api.typed.ReturningType
-import pl.touk.nussknacker.engine.avro.{KafkaAvroFactory, KafkaAvroSpecMixin}
+import pl.touk.nussknacker.engine.api.typed.typing.{Typed, TypedObjectTypingResult, Unknown}
+import pl.touk.nussknacker.engine.api.{MetaData, StreamMetaData}
+import pl.touk.nussknacker.engine.avro.KafkaAvroFactory.{SchemaVersionParamName, TopicParamName}
 import pl.touk.nussknacker.engine.avro.schema.{FullNameV1, FullNameV2}
 import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.ConfluentSchemaRegistryProvider
 import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.client._
 import pl.touk.nussknacker.engine.avro.schemaregistry.{SchemaSubjectNotFound, SchemaVersionNotFound}
 import pl.touk.nussknacker.engine.avro.typed.AvroSchemaTypeDefinitionExtractor
+import pl.touk.nussknacker.engine.avro.{KafkaAvroFactory, KafkaAvroSpecMixin}
+import pl.touk.nussknacker.engine.compile.ExpressionCompiler
+import pl.touk.nussknacker.engine.compile.nodevalidation.{GenericNodeTransformationValidator, TransformationResult}
+import pl.touk.nussknacker.engine.expression.ExpressionEvaluator
+import pl.touk.nussknacker.engine.graph.evaluatedparam.Parameter
+import pl.touk.nussknacker.engine.graph.expression.Expression
+import pl.touk.nussknacker.engine.spel.Implicits._
+import pl.touk.nussknacker.engine.testing.{EmptyProcessConfigCreator, LocalModelData}
 
 class KafkaAvroSourceFactorySpec extends KafkaAvroSpecMixin with KafkaAvroSourceSpecMixin {
 
@@ -60,7 +74,7 @@ class KafkaAvroSourceFactorySpec extends KafkaAvroSpecMixin with KafkaAvroSource
     val sourceFactory = createAvroSourceFactory(useSpecificAvroReader = false)
     val givenObj = FullNameV2.createRecord("Jan", "Maria", "Kowalski")
 
-    assertThrowsWithParent[SchemaVersionNotFound]{
+    assertThrowsWithParent[SchemaVersionNotFound] {
       readLastMessageAndVerify(sourceFactory, RecordTopic, 3, givenObj, FullNameV2.schema)
     }
   }
@@ -93,6 +107,63 @@ class KafkaAvroSourceFactorySpec extends KafkaAvroSpecMixin with KafkaAvroSource
   protected def createAvroSourceFactory(useSpecificAvroReader: Boolean): KafkaAvroSourceFactory[GenericRecord] = {
     val schemaRegistryProvider = createSchemaRegistryProvider[GenericRecord](useSpecificAvroReader)
     new KafkaAvroSourceFactory(schemaRegistryProvider, processObjectDependencies, None)
+  }
+
+  test("Should validate specific version") {
+    val result = validate(TopicParamName -> s"'${KafkaAvroSourceMockSchemaRegistry.RecordTopic}'",
+      SchemaVersionParamName -> "1")
+
+    result.errors shouldBe Nil
+  }
+
+  test("Should validate latest version") {
+    val result = validate(TopicParamName -> s"'${KafkaAvroSourceMockSchemaRegistry.RecordTopic}'",
+      SchemaVersionParamName -> "")
+
+    result.errors shouldBe Nil
+  }
+
+  test("Should return sane error on invalid topic") {
+    val result = validate(TopicParamName -> "'terefere'", SchemaVersionParamName -> "")
+
+    result.errors shouldBe CustomNodeError("id", "Schema subject doesn't exist.", Some(TopicParamName)) ::
+      CustomNodeError("id", "Schema subject doesn't exist.", Some(SchemaVersionParamName)) :: Nil
+    result.outputContext shouldBe ValidationContext(Map(Interpreter.InputParamName -> Unknown))
+  }
+
+  test("Should return sane error on invalid version") {
+    val result = validate(TopicParamName -> s"'${KafkaAvroSourceMockSchemaRegistry.RecordTopic}'",
+      SchemaVersionParamName -> "12345")
+
+    result.errors shouldBe CustomNodeError("id", "Schema version doesn't exist.", Some(SchemaVersionParamName)) :: Nil
+    result.outputContext shouldBe ValidationContext(Map(Interpreter.InputParamName -> Unknown))
+  }
+
+  test("Should properly detect input type") {
+    val result = validate(TopicParamName -> s"'${KafkaAvroSourceMockSchemaRegistry.RecordTopic}'",
+      SchemaVersionParamName -> "")
+
+    result.errors shouldBe Nil
+    result.outputContext shouldBe ValidationContext(Map(Interpreter.InputParamName -> TypedObjectTypingResult(
+      Map(
+        "first" -> Typed[CharSequence],
+        "middle" -> Typed[CharSequence],
+        "last" -> Typed[CharSequence]
+      ), Typed.typedClass[GenericRecord]
+    )))
+  }
+
+  private def validate(params: (String, Expression)*): TransformationResult = {
+
+    val modelData = LocalModelData(ConfigFactory.empty(), new EmptyProcessConfigCreator)
+
+    val validator = new GenericNodeTransformationValidator(ExpressionCompiler.withoutOptimization(modelData),
+      ExpressionEvaluator.unOptimizedEvaluator(modelData))
+
+    implicit val meta: MetaData = MetaData("processId", StreamMetaData())
+    implicit val nodeId: NodeId = NodeId("id")
+    val paramsList = params.toList.map(p => Parameter(p._1, p._2))
+    validator.validateNode(createAvroSourceFactory(false), paramsList, ValidationContext(), Some(Interpreter.InputParamName)).toOption.get
   }
 
   private def createKeyValueAvroSourceFactory[K: TypeInformation, V: TypeInformation]: KafkaAvroSourceFactory[(K, V)] = {
